requirements.txt
Flask
langchain
langchain-google-genai
python-dotenv